调整API名称和方法
